Remove LDVerifierKeyPair, rename export helper functions.
