+++ lib/LDKeyPair.js
@@ -66,6 +66,9 @@ class LDKeyPair {
    * Exports the serialized representation of the KeyPair
    * and other information that json-ld Signatures can use to form a proof.
    *
+   * NOTE: Subclasses MUST override this method (and add the exporting of
+   * their public and private key material).
+   *
    * @param {boolean} [publicKey] - Export public key material?
    * @param {boolean} [privateKey] - Export private key material?
    *
@@ -82,34 +85,8 @@ class LDKeyPair {
       type: this.type,
       controller: this.controller
     };
-    if(publicKey) {
-      this.addPublicKey({key}); // Subclass-specific
-    }
-    if(privateKey) {
-      this.addPrivateKey({key}); // Subclass-specific
-    }
+
     return key;
-  }
-
-  /**
-   * Adds the suite-specific public key material, serialized to string, to
-   * the exported public key node.
-   * @param {object} key - Public key object.
-   * @returns {object}
-   */
-  addPublicKey(/* {key} */) {
-    throw new Error('Abstract method, must be implemented in subclass.');
-  }
-
-  /**
-   * Adds the suite-specific private key material, serialized to string, to
-   * the exported public key node.
-   *
-   * @param {object} key - Private key object.
-   * @returns {object}
-   */
-  addPrivateKey(/* {key} */) {
-    throw new Error('Abstract method, must be implemented in subclass.');
   }
 
   /**
@@ -137,6 +114,52 @@ class LDKeyPair {
    */
   verifyFingerprint(/* {fingerprint} */) {
     throw new Error('Abstract method, must be implemented in subclass.');
+  }
+
+  /* eslint-disable max-len */
+  /**
+   * Returns a signer object for use with
+   * [jsonld-signatures]{@link https://github.com/digitalbazaar/jsonld-signatures}.
+   * NOTE: Applies only to verifier type keys (like ed25519).
+   *
+   * @example
+   * > const signer = keyPair.signer();
+   * > signer
+   * { sign: [AsyncFunction: sign] }
+   * > signer.sign({data});
+   *
+   * @returns {{sign: Function}} A signer for json-ld usage.
+   */
+  /* eslint-enable */
+  signer() {
+    return {
+      async sign({/* data */}) {
+        throw new Error('Abstract method, must be implemented in subclass.');
+      }
+    };
+  }
+
+  /* eslint-disable max-len */
+  /**
+   * Returns a verifier object for use with
+   * [jsonld-signatures]{@link https://github.com/digitalbazaar/jsonld-signatures}.
+   * NOTE: Applies only to verifier type keys (like ed25519).
+   *
+   * @example
+   * > const verifier = keyPair.verifier();
+   * > verifier
+   * { verify: [AsyncFunction: verify] }
+   * > verifier.verify(key);
+   *
+   * @returns {{verify: Function}} Used to verify jsonld-signatures.
+   */
+  /* eslint-enable */
+  verifier() {
+    return {
+      async verify({/* data, signature */}) {
+        throw new Error('Abstract method, must be implemented in subclass.');
+      }
+    };
   }
 }
