+++ lib/index.js
@@ -5,10 +5,8 @@
 
 const {CryptoLD} = require('./CryptoLD');
 const {LDKeyPair} = require('./LDKeyPair');
-const {LDVerifierKeyPair} = require('./LDVerifierKeyPair');
 
 module.exports = {
   CryptoLD,
-  LDKeyPair,
-  LDVerifierKeyPair
+  LDKeyPair
 };
